chore: api directory move to src (#16)
